requirements.txt
playwright==1.29.1
pytest-playwright==0.3.0
python-a11y-playwright==1.0.1